backend/requirements.txt
fastapi
uvicorn
gunicorn
pydantic
pydantic[email]
python-jose
passlib[bcrypt]
pymongo
pytest
httpx
mongomock
python-multipart
python-dotenv
